fix: use c_ulong for ioctl request constants on Linux
Linux libc::ioctl takes c_ulong (u64) for the request parameter,
not c_int. The previous commit had these typed incorrectly.

+++ crates/halt-sandbox/src/linux.rs
@@ -232,8 +232,8 @@ pub fn apply_landlock(_ruleset: ()) -> Result<(), SandboxError> {
 #[cfg(target_os = "linux")]
 unsafe fn bring_up_loopback() {
     const IFNAMSIZ: usize = 16;
-    const SIOCGIFFLAGS: libc::c_int = 0x8913;
-    const SIOCSIFFLAGS: libc::c_int = 0x8914;
+    const SIOCGIFFLAGS: libc::c_ulong = 0x8913;
+    const SIOCSIFFLAGS: libc::c_ulong = 0x8914;
     const IFF_UP: i16 = 0x1;
 
     /// Mirrors the C `struct ifreq` layout for SIOCGIFFLAGS / SIOCSIFFLAGS.
